backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
pydantic = "^1.10.9"
sqlalchemy = "^2.0.0"
mysql-connector-python = "^8.0.33"
sqlmodel = "^0.0.22"
pymysql = "^1.1.1"
cryptography = "^43.0.3"
fastapi-cors = "^0.0.6"
google-auth = "^2.35.0"
pulp = "^2.9.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
mangum = "^0.19.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"